+++ components/segmenter/src/complex/dictionary.rs
@@ -3,14 +3,12 @@
 // (online at: https://github.com/unicode-org/icu4x/blob/main/LICENSE ).
 
 use crate::grapheme::*;
-use crate::indices::Utf16Indices;
+use crate::indices::*;
 use crate::provider::*;
 #[cfg(feature = "unstable")]
 use crate::scaffold::PotentiallyIllFormedUtf8;
 use crate::scaffold::{RuleBreakType, Utf8, Utf16};
 use icu_collections::char16trie::{Char16Trie, TrieResult};
-#[cfg(feature = "unstable")]
-use utf8_iter::Utf8CharIndices;
 
 /// Lifetimes:
 /// - `'data` = lifetime of the data
