+++ components/segmenter/src/provider/complex/dictionary.rs
@@ -1,0 +1,34 @@
+// This file is part of ICU4X. For terms of use, please see the file
+// called LICENSE at the top level of the ICU4X source tree
+// (online at: https://github.com/unicode-org/icu4x/blob/main/LICENSE ).
+
+use icu_provider::prelude::*;
+use zerovec::ZeroVec;
+
+/// char16trie data for dictionary break
+///
+/// <div class="stab unstable">
+/// 🚧 This code is considered unstable; it may change at any time, in breaking or non-breaking ways,
+/// including in SemVer minor releases. While the serde representation of data structs is guaranteed
+/// to be stable, their Rust representation might not be. Use with caution.
+/// </div>
+#[derive(Debug, PartialEq, Clone, yoke::Yokeable, zerofrom::ZeroFrom)]
+#[cfg_attr(feature = "datagen", derive(serde::Serialize, databake::Bake))]
+#[cfg_attr(feature = "datagen", databake(path = icu_segmenter::provider))]
+#[cfg_attr(feature = "serde", derive(serde::Deserialize))]
+pub struct UCharDictionaryBreakData<'data> {
+    /// Dictionary data of char16trie.
+    #[cfg_attr(feature = "serde", serde(borrow))]
+    pub trie_data: ZeroVec<'data, u16>,
+}
+
+icu_provider::data_struct!(
+    UCharDictionaryBreakData<'_>,
+    #[cfg(feature = "datagen")]
+);
+
+pub(crate) struct UCharDictionaryBreakDataV1;
+
+impl DynamicDataMarker for UCharDictionaryBreakDataV1 {
+    type DataStruct = UCharDictionaryBreakData<'static>;
+}
